Thinko in vms.c:copy_expand_unix_filename_escape().
Turns out comparing a signed byte to values above 0x7f doesn't make
sense.  *All* signed byte integers are less than or equal to 0x9f,
so the other two branches of the if could never be taken.

This code probably needs more review and testing, but we might as
well make it do what it intends to do before reviewing those
intentions and factoring out some of the copy-and-paste verbosity.