Move the MUTLIARCH macro soup for setting BYTEORDER from config.h to perl.h
MUTLIARCH and CROSS_COMPILE ignore the probed byteorder, instead using C
compiler macros to determine the byteorder for the currently compiling
architecture. The first implementation of this logic dates back to
40750cc09a264b13 in 1996, which added pre-processor code to config_h.SH
for NeXT only. Subsequent changes have made the code more "portable" by
keying it off "MUTLIARCH" instead of "NeXT", and accounting for 32 and 64
bit word sizes.

However, as the code is itself decision logic, not simply conveying the
results of probing, it doesn't really belong in config.h. perl.h feels a
more logical place for it.