Remove 3 unused scripts from Porting
Porting/findvars was added in 1998 in 2bd2b9e04a68ec86. It searches @ARGV for
its wordlist of then-current interpreter variable names.
Porting/fixvars was added as fixvars in 1998 in a15299417de39f35. It captures
the output of make (defaulting to make miniperl), parses it for errors
matching /undeclared/, and then attempts to edit the relevant line of the
reported file to prefix the name with PL_
Porting/fixCORE is a modified copy of Porting/fixvars, added in 1998 in
a8693bd382efcc6d. It's intended to load modules, catching errors of the form
/Ambiguous call resolved as CORE::/, and editing the relevant lines to prefix
CORE:: to the function in question. It appears only ever to have been used to
fix warnings in Math::Complex.

All 3 have only had trivial style and spelling edits since addition, and
have been unused for over 12 years.