In POSIX, drastically simplify the wrappers for "unimplemented" functions.
Replace all the subroutines that croak() with a data structure and 8 lines in
POSIX::AUTOLOAD().

[By Aristotle Pagaltzis, with some editing by the committer, and most of his
message changes applied as a previous commit to split apart improvements from
pure refactoring]

This commit eliminates the helper functions POSIX::refef() and
POSIX::unimpl(), which were not part of the documented API, not exported,
and not used in any code outside the core (that is visible to Google
codesearch).