utf8_heavy: Correctly handle \p{L_}
L_ is an old, discouraged but not deprecated,  synonym for LC, cased
letters.  It was improperly being converted to simply L, all letters,
because it is supposed to match loosely, and the trailing underscore was
being stripped.  It needs a special case.

Tests passed for this, as the machine generated test case happened to
choose code points that are in both L and LC.  Also some tests were
testing L instead of LC because of a similar flaw.