Make (l)stat respect get-magic on globs and globrefs
They were ignoring get-magic for those.

A side effect of this fix is that lstat filehandle warnings and errors
are now consistent:

lstat _ used to die if the previous stat was not an lstat, but
lstat *_ and lstat \*_ would happily return what was in the buffer.
Now they die.

lstat FH and \*FH used to warn, but not lstat *FH.  Now it does.

See bug #98864.