[perl #92436] Make Filter::Simple match variables better
(This is the second bug reported in ticket #92436.)

Filter::Simple was using Text::Balanced’s extract_variable, which

       ...extracts any valid Perl variable or variable-
       involved expression, including scalars, arrays, hashes, array
       accesses, hash look-ups, method calls through objects, subrou-
       tine calls through subroutine references, etc.

So it extracts things like $x->foo("blah lbah blah").  That means
that, when the user of Filter::Simple asks for everything except
strings, "blah lbah blah" is passed through anyway, because
Filter::Simple things it’s part of a variable name.  That obviously
doesn’t work.

This commit makes Filter::Simple use a regular expression for varia-
bles, as it does for other things.  It’s certainly not foolproof, but
Filter::Simple in general is not foolproof, and this regular expres-
sion is actually less foolproof than most parts of Filter::Simple.  So
it’s a step in the right direction (unless you consider deletion to be
the right direction).