Add a --regen option to checkcfgvar.pl to regenerate config files.
Verify that the section of config file containing probed files is sorted
lexically. If --regen is used, updated the file on disk with a correctly
sorted version. (Except for configure.com, which has a different structure
not amenable to automatic analysis and update, hence still has to be
updated manually.)

Ensure all config files are correctly sorted.