[perl #93638] $ENV{LS_COLORS} causes miniperl glob failure
On some systems, csh croaks if the LS_COLORS envvar contains something
it considers invalid.

The easiest and least intrusive fix for now is to localize %ENV before
running csh, though eventually it might be nice to use File::Glob in
miniperl, either by linking it statically or by inlining it.