base.pm no longer tracks loaded modules
Tracking modules loaded by base.pm to avoid reloading them means
that any module that wishes to force a reload can't just modify
%INC but must also interact with base.pm to clear its cache.

Removing the module-loaded checks have minimal impact.  The internal
call to require() is a NOP for a loaded module.  For an internal
package, the require() will fail with the same error message as
before, which base.pm then ignores.  Tracking modules loaded to avoid
this slight extra overhead is not worth the complexity for other
modules that wish to manipulate %INC.