handy.h: Simplify isASCII definition
Thus retains essentially the same definition for EBCDIC platforms, but
substitutes a simpler one for ASCII platforms.  On my system, the new
definition compiles to about half the assembly instructions that the old
one did (non-optimized)

A bomb-proof definition of ASCII is to make sure that the value is
unsigned in the largest possible unsigned for the platform so there is
no possible loss of information, and then the ord  must be < 128.