Don't use swash to find cntrls
Unicode stability policy guarantees that no code points will ever be
added to the control characters beyond those already in it.

All such characters are in the Latin1 range, and so the Perl core
already knows which ones those are, and so there is no need to go out to
disk and create a swash for these.