regexec.c: Avoid hard-coded utf8 tests for EBCDIC
When a swash is loaded, generally it is checked for sanity with an
assert().  The strings used are hard-coded utf8 strings, which will be
different in EBCDIC, and hence will fail.  I haven't figured out a
simple way to get compile-time utf8 vs utfebcdic strings, but we can
just skip the check in EBCDIC builds