regcomp.c: Add invlist_invert_prop()
This new function inverts a Unicode property.  A regular inversion
doesn't work because it operates on the whole of the code space, and
Unicode property inversions don't invert above-Unicode code points.

This does for inversion lists, what an earlier commit did for swashes.
This function is currently not called by anyone.