To build perl-5.003 bisect-runner.pl needs to force a value for trnl.
Without this the current makedepend will hang waiting on stdin.
Additionally, when bisecting this far back we need to watch out for the
//depot/perlext/Compiler branch, which has no Configure, patchlevel.h or
other perl source, but may be checked out by the bisect, as it is the second
parent of a8581515f26a081f. If the file Configure is missing, we exit 125
to signal a skip.