[perl #99984] Incorrect errmsg with our $::é
Having PL_parser->error_count set to non-zero when utf8_heavy.pl tries
to do() one of its swashes results in ‘Compilation error’ being placed
in $@ during the do, even if it was successful.  This patch sets the
error_count to 0 before calling SWASHNEW, to prevent that.  It uses
SAVEI8, to make sure it is restored on scope exit.