handy.h: Reorder tests for speed
It's much more likely that a random character will have its ordinal be
above the ordinal for '7' than below.  In the test for if a character is
octal then, testing first if it is <= '7' will exclude many more
possibilities than if the first test is if it is >= '0'.

I left the ones for lowercase letters in the same order, because, in
ASCII, anyway, there are more characters below 'a' than above it.